Changes in stratum to rename from ES2000 to ES2K
