+++ stratum/hal/bin/tdi/es2k/es2k_skip_p4.conf
