+++ stratum/hal/lib/tdi/es2k/es2k_chassis_manager.h
@@ -2,8 +2,8 @@
 // Copyright 2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef STRATUM_HAL_LIB_TDI_ES2000_ES2K_CHASSIS_MANAGER_H_
-#define STRATUM_HAL_LIB_TDI_ES2000_ES2K_CHASSIS_MANAGER_H_
+#ifndef STRATUM_HAL_LIB_TDI_ES2K_CHASSIS_MANAGER_H_
+#define STRATUM_HAL_LIB_TDI_ES2K_CHASSIS_MANAGER_H_
 
 #include <map>
 #include <memory>
@@ -15,7 +15,7 @@
 #include "absl/types/optional.h"
 #include "stratum/glue/integral_types.h"
 #include "stratum/hal/lib/tdi/tdi_sde_interface.h"
-#include "stratum/hal/lib/tdi/es2000/es2k_port_manager.h"
+#include "stratum/hal/lib/tdi/es2k/es2k_port_manager.h"
 #include "stratum/hal/lib/common/gnmi_events.h"
 #include "stratum/hal/lib/common/phal_interface.h"
 #include "stratum/hal/lib/common/utils.h"
@@ -267,4 +267,4 @@ class Es2kChassisManager {
 }  // namespace hal
 }  // namespace stratum
 
-#endif  // STRATUM_HAL_LIB_TDI_ES2000_ES2K_CHASSIS_MANAGER_H_
+#endif  // STRATUM_HAL_LIB_TDI_ES2K_CHASSIS_MANAGER_H_
