+++ stratum/hal/lib/tdi/es2k/es2k_hal.cc
@@ -3,7 +3,7 @@
 // Copyright 2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "stratum/hal/lib/tdi/es2000/es2k_hal.h"
+#include "stratum/hal/lib/tdi/es2k/es2k_hal.h"
 
 #include <limits.h>
 #include <utility>
