+++ stratum/hal/lib/tdi/es2k/es2k_sde_utils.cc
