+++ stratum/hal/lib/tdi/es2k/es2k_switch.h
@@ -2,8 +2,8 @@
 // Copyright 2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef STRATUM_HAL_LIB_TDI_ES2000_ES2K_SWITCH_H_
-#define STRATUM_HAL_LIB_TDI_ES2000_ES2K_SWITCH_H_
+#ifndef STRATUM_HAL_LIB_TDI_ES2K_SWITCH_H_
+#define STRATUM_HAL_LIB_TDI_ES2K_SWITCH_H_
 
 #include <map>
 #include <memory>
@@ -12,7 +12,7 @@
 
 #include "absl/synchronization/mutex.h"
 #include "stratum/hal/lib/tdi/tdi_node.h"
-#include "stratum/hal/lib/tdi/es2000/es2k_chassis_manager.h"
+#include "stratum/hal/lib/tdi/es2k/es2k_chassis_manager.h"
 #include "stratum/hal/lib/common/switch_interface.h"
 
 // Suppress clang errors
@@ -125,4 +125,4 @@ class Es2kSwitch : public SwitchInterface {
 }  // namespace hal
 }  // namespace stratum
 
-#endif  // STRATUM_HAL_LIB_TDI_ES2000_ES2K_SWITCH_H_
+#endif  // STRATUM_HAL_LIB_TDI_ES2K_SWITCH_H_
