+++ stratum/hal/lib/tdi/es2k/test_main.cc
@@ -19,7 +19,7 @@ int main(int argc, char** argv) {
 
   bool tmpdir_created = false;
   if (FLAGS_test_tmpdir.empty()) {
-    char tmpdir[] = "/tmp/stratum_hal_tdi_es2000_test.XXXXXX";
+    char tmpdir[] = "/tmp/stratum_hal_tdi_es2k_test.XXXXXX";
     CHECK(mkdtemp(tmpdir));
     FLAGS_test_tmpdir = tmpdir;
     tmpdir_created = true;
